feat(network): implement basic peer banning (#6659)
Description
---
feat(network): implement basic peer banning
feat(network): basic impl of fail2ban logs
fix: improve whois command

Motivation and Context
---
Peers were marked as banned but the ban was not enforced.

This PR also includes logs that can be used to implement fail2ban.

Note bans are not enforced in the application on restart (no
persistence). However if fail2ban is implemented the ban will continue
to be enforced there.

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify